fix: font issue w build
