+++ frontend/src/app/fonts.ts
@@ -17,6 +17,20 @@ export const workSans = Work_Sans({
   variable: "--font-work-sans",
 });
 
+export const gentona = localFont({
+  src: [
+    { path: "./fonts/gentona/Gentona Thin.otf", weight: "100" },
+    { path: "./fonts/gentona/Gentona ExtraLight.otf", weight: "200" },
+    { path: "./fonts/gentona/Gentona Light.otf", weight: "300" },
+    { path: "./fonts/gentona/Gentona Book.otf", weight: "400" },
+    { path: "./fonts/gentona/Gentona Medium.otf", weight: "500" },
+    { path: "./fonts/gentona/Gentona SemiBold.otf", weight: "600" },
+    { path: "./fonts/gentona/Gentona Bold.otf", weight: "700" },
+    { path: "./fonts/gentona/Gentona ExtraBold.otf", weight: "800" },
+    { path: "./fonts/gentona/Gentona Heavy.otf", weight: "900" },
+  ],
+});
+
 export const gentonaMedium = localFont({
   src: "../../public/fonts/Gentona Medium.otf",
   display: "swap",
